move convert to push to pubsub.go
