+++ pkg/pubsub/adapter/converters/pubsub.go
@@ -18,11 +18,15 @@ package converters
 
 import (
 	"context"
+	"encoding/json"
+	"time"
 
 	cloudevents "github.com/cloudevents/sdk-go"
 	. "github.com/cloudevents/sdk-go/pkg/cloudevents"
 	cepubsub "github.com/cloudevents/sdk-go/pkg/cloudevents/transport/pubsub"
 	pubsubcontext "github.com/cloudevents/sdk-go/pkg/cloudevents/transport/pubsub/context"
+	"go.uber.org/zap"
+	"knative.dev/pkg/logging"
 
 	"github.com/google/knative-gcp/pkg/apis/events/v1alpha1"
 )
@@ -48,6 +52,8 @@ func convertPubSub(ctx context.Context, msg *cepubsub.Message, sendMode ModeType
 	event.Data = msg.Data
 	event.DataEncoded = true
 	// Attributes are extensions.
+	// Also grab all extensions as a string, which we will set as attributes payload if send mode is Push.
+	attrs := make(map[string]string, 0)
 	if msg.Attributes != nil && len(msg.Attributes) > 0 {
 		for k, v := range msg.Attributes {
 			// CloudEvents v1.0 attributes MUST consist of lower-case letters ('a' to 'z') or digits ('0' to '9') as per
@@ -56,7 +62,76 @@ func convertPubSub(ctx context.Context, msg *cepubsub.Message, sendMode ModeType
 			if IsAlphaNumeric(k) {
 				event.SetExtension(k, v)
 			}
+			// All atributes (whether valid as extensions) are stored in attrs to be set in the payload if mode is Push.
+			attrs[k] = v
 		}
 	}
+	// If send mode is Push, convert to Pub/Sub Push payload style.
+	if sendMode == Push {
+		event = convertToPush(ctx, event, attrs)
+	}
 	return &event, nil
+}
+
+// PushMessage represents the format Pub/Sub uses to push events.
+type PushMessage struct {
+	// Subscription is the subscription ID that received this Message.
+	Subscription string `json:"subscription"`
+	// Message holds the Pub/Sub message contents.
+	Message *PubSubMessage `json:"message,omitempty"`
+}
+
+// PubSubMessage matches the inner message format used by Push Subscriptions.
+type PubSubMessage struct {
+	// ID identifies this message. This ID is assigned by the server and is
+	// populated for Messages obtained from a subscription.
+	// This field is read-only.
+	ID string `json:"id,omitempty"`
+
+	// Data is the actual data in the message.
+	Data interface{} `json:"data,omitempty"`
+
+	// Attributes represents the key-value pairs the current message
+	// is labelled with.
+	Attributes map[string]string `json:"attributes,omitempty"`
+
+	// The time at which the message was published. This is populated by the
+	// server for Messages obtained from a subscription.
+	// This field is read-only.
+	PublishTime time.Time `json:"publish_time,omitempty"`
+}
+
+// convertToPush convert an event to a Pub/Sub style Push payload.
+func convertToPush(ctx context.Context, event cloudevents.Event, attrs map[string]string) cloudevents.Event {
+	logger := logging.FromContext(ctx).With(zap.Any("event.id", event.ID()))
+
+	tx := pubsubcontext.TransportContextFrom(ctx)
+
+	push := cloudevents.NewEvent(event.SpecVersion())
+	push.Context = event.Context.Clone()
+
+	msg := &PubSubMessage{
+		ID:          event.ID(),
+		Attributes:  attrs,
+		PublishTime: event.Time(),
+	}
+
+	var raw json.RawMessage
+	if err := event.DataAs(&raw); err != nil {
+		logger.Debugw("Failed to get data as raw json, using as is.", zap.Error(err))
+		// Use data as a byte slice.
+		msg.Data = event.Data
+	} else {
+		// Use data as a raw message.
+		msg.Data = raw
+	}
+
+	if err := push.SetData(&PushMessage{
+		Subscription: tx.Subscription,
+		Message:      msg,
+	}); err != nil {
+		logger.Warnw("Failed to set data.", zap.Error(err))
+	}
+
+	return push
 }
