+++ pkg/pubsub/adapter/converters/pubsub_test.go
@@ -126,3 +126,94 @@ func pubSubCloudEvent(extensions map[string]string) *cloudevents.Event {
 	}
 	return &e
 }
+
+func TestConvertToPush_noattrs(t *testing.T) {
+	want := `Validation: valid
+Context Attributes,
+  specversion: 0.2
+  type: unit.testing
+  source: source
+  id: abc-123
+  contenttype: application/json
+Data,
+  {
+    "subscription": "sub",
+    "message": {
+      "id": "abc-123",
+      "data": "testing",
+      "publish_time": "0001-01-01T00:00:00Z"
+    }
+  }
+`
+
+	event := cloudevents.NewEvent()
+	event.SetSource("source")
+	event.SetType("unit.testing")
+	event.SetID("abc-123")
+	event.SetDataContentType("application/json")
+	_ = event.SetData("testing")
+	attrs := make(map[string]string, 0)
+	ctx := pubsubcontext.WithTransportContext(context.TODO(), pubsubcontext.NewTransportContext(
+		"proj",
+		"top",
+		"sub",
+		"test",
+		&pubsub.Message{},
+	))
+
+	got := convertToPush(ctx, event, attrs)
+
+	if diff := cmp.Diff(want, got.String()); diff != "" {
+		t.Logf("%s", got.String())
+		t.Errorf("failed to get expected (-want, +got) = %v", diff)
+	}
+}
+
+func TestConvertToPush_attrs(t *testing.T) {
+	want := `Validation: valid
+Context Attributes,
+  specversion: 0.2
+  type: unit.testing
+  source: source
+  id: abc-123
+  contenttype: application/json
+Extensions,
+  foo: bar
+Data,
+  {
+    "subscription": "sub",
+    "message": {
+      "id": "abc-123",
+      "data": "testing",
+      "attributes": {
+        "foo": "bar"
+      },
+      "publish_time": "0001-01-01T00:00:00Z"
+    }
+  }
+`
+
+	event := cloudevents.NewEvent()
+	event.SetSource("source")
+	event.SetType("unit.testing")
+	event.SetID("abc-123")
+	event.SetExtension("foo", "bar")
+	event.SetDataContentType("application/json")
+	_ = event.SetData("testing")
+	attrs := make(map[string]string, 0)
+	attrs["foo"] = "bar"
+	ctx := pubsubcontext.WithTransportContext(context.TODO(), pubsubcontext.NewTransportContext(
+		"proj",
+		"top",
+		"sub",
+		"test",
+		&pubsub.Message{},
+	))
+
+	got := convertToPush(ctx, event, attrs)
+
+	if diff := cmp.Diff(want, got.String()); diff != "" {
+		t.Logf("%s", got.String())
+		t.Errorf("failed to get expected (-want, +got) = %v", diff)
+	}
+}
